image_scraper/requirements.txt
streamlit
requests
beautifulsoup4
selenium
webdriver-manager
Pillow
